Update swagger AnnotationsStudentDetail add format="uuid" Created Home FilterLists folder in Swagger and moved ```AnnotationsSpecializationList```& ``` AnnotationsDevelopmentList```files into it. Taken from the ‘student/resume’ group these two paths: ‘development/list & specialization/list’.
