+++ app/Annotations/OpenApi/controllersAnnotations/homeFilterLists/AnnotationsSpecializationList.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Annotations\OpenApi\controllersAnnotations\specializationListAnnotations;
+namespace App\Annotations\OpenApi\controllersAnnotations\homeFilterLists;
 
 use OpenApi\Annotations as OA;
 
@@ -10,7 +10,7 @@ class AnnotationsSpecializationList
      * @OA\Get(
      *     path="/specialization/list",
      *     summary="Retrieve a specialization list from resume model enum",
-     *     tags={"Specialization"},
+     *     tags={"Home Filter Lists"},
      *     @OA\Response(
      *         response=200,
      *         description="Successful. Specialization list retrieved from enum",
