+++ storage/api-docs/api-docs.json
@@ -532,13 +532,37 @@
         "/development/list": {
             "get": {
                 "tags": [
-                    "Development"
+                    "Home Filter Lists"
                 ],
                 "summary": "Retrieve a development list ",
-                "operationId": "e14a9caf5611cd7fee512d6251dbc24f",
+                "operationId": "316ec2428e9951c98d27a1cde1db075c",
                 "responses": {
                     "200": {
                         "description": "Successful. Development list retrieved from json file",
+                        "content": {
+                            "application/json": {
+                                "schema": {
+                                    "type": "array",
+                                    "items": {
+                                        "type": "string"
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        },
+        "/specialization/list": {
+            "get": {
+                "tags": [
+                    "Home Filter Lists"
+                ],
+                "summary": "Retrieve a specialization list from resume model enum",
+                "operationId": "2fef70eb5bb38db63cd2428637e229cf",
+                "responses": {
+                    "200": {
+                        "description": "Successful. Specialization list retrieved from enum",
                         "content": {
                             "application/json": {
                                 "schema": {
@@ -928,30 +952,6 @@
                         "bearerAuth": []
                     }
                 ]
-            }
-        },
-        "/specialization/list": {
-            "get": {
-                "tags": [
-                    "Specialization"
-                ],
-                "summary": "Retrieve a specialization list from resume model enum",
-                "operationId": "ed66866e7488b350aa0142e3b004ac52",
-                "responses": {
-                    "200": {
-                        "description": "Successful. Specialization list retrieved from enum",
-                        "content": {
-                            "application/json": {
-                                "schema": {
-                                    "type": "array",
-                                    "items": {
-                                        "type": "string"
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
             }
         },
         "/students": {
@@ -1769,7 +1769,8 @@
                         "description": "Student ID",
                         "required": true,
                         "schema": {
-                            "type": "string"
+                            "type": "string",
+                            "format": "uuid"
                         }
                     }
                 ],
